+++ packages/react/src/hooks/index.ts
@@ -5,3 +5,4 @@ export * from './use-nfts';
 export * from './use-encrypted-content';
 export * from './use-download-content';
 export * from './use-serverside-url';
+export * from './use-is-ios';