+++ packages/react/src/hooks/use-is-ios.tsx
@@ -1,0 +1,7 @@
+import { useMemo } from 'react';
+
+export const useIsIOSWithHlsSupport = () => {
+  return useMemo(() => {
+    return /^((?!chrome|android).)*safari/i.test(navigator.userAgent);
+  }, []);
+};